+++ src/apps/viewer/index.ts
@@ -8,5 +8,5 @@
 import './embedded.html';
 import './favicon.ico';
 import './index.html';
-require('mol-plugin-ui/skin/light.scss');
+import '../../mol-plugin-ui/skin/light.scss';
 export * from './app';