+++ src/mol-gl/shader/chunks/common-frag-params.glsl.ts
@@ -15,6 +15,7 @@ uniform vec4 uLod;
     uniform vec3 uClipObjectPosition[dClipObjectCount];
     uniform vec4 uClipObjectRotation[dClipObjectCount];
     uniform vec3 uClipObjectScale[dClipObjectCount];
+    uniform mat4 uClipObjectTransform[dClipObjectCount];
 
     #if defined(dClipping)
         #if __VERSION__ == 100 || defined(dClippingType_instance) || !defined(dVaryingGroup)