+++ src/mol-repr/volume/slice.ts
@@ -18,20 +18,239 @@ import { PickingId } from '../../mol-geo/geometry/picking';
 import { EmptyLoci, Loci } from '../../mol-model/loci';
 import { Interval, SortedArray } from '../../mol-data/int';
 import { transformPositionArray } from '../../mol-geo/util';
-import { RenderableState } from '../../mol-gl/renderable';
 import { Color } from '../../mol-util/color';
 import { ColorTheme } from '../../mol-theme/color';
 import { packIntToRGBArray } from '../../mol-util/number-packing';
 import { eachVolumeLoci } from './util';
 import { Vec3 } from '../../mol-math/linear-algebra/3d/vec3';
-import { equalEps } from '../../mol-math/linear-algebra/3d/common';
+import { Quat } from '../../mol-math/linear-algebra/3d/quat';
+import { degToRad } from '../../mol-math/misc';
+import { Mat4 } from '../../mol-math/linear-algebra/3d/mat4';
+import { clamp, normalize } from '../../mol-math/interpolate';
 
-export async function createImage(ctx: VisualContext, volume: Volume, key: number, theme: Theme, props: PD.Values<SliceParams>, image?: Image) {
+export const SliceParams = {
+    ...Image.Params,
+    quality: { ...Image.Params.quality, isEssential: false },
+    dimension: PD.MappedStatic('x', {
+        x: PD.Numeric(0, { min: 0, max: 0, step: 1 }, { immediateUpdate: true }),
+        y: PD.Numeric(0, { min: 0, max: 0, step: 1 }, { immediateUpdate: true }),
+        z: PD.Numeric(0, { min: 0, max: 0, step: 1 }, { immediateUpdate: true }),
+    }, { isEssential: true, hideIf: p => p.mode !== 'grid', description: 'Slice position in grid coordinates.' }),
+    isoValue: Volume.IsoValueParam,
+    mode: PD.Select('grid', PD.arrayToOptions(['grid', 'frame'] as const), { description: 'Grid: slice through the volume along the grid axes in integer steps. Frame: slice through the volume along arbitrary axes in any step size.' }),
+    offset: PD.Numeric(0, { min: -1, max: 1, step: 0.01 }, { isEssential: true, immediateUpdate: true, hideIf: p => p.mode !== 'frame', description: 'Relative offset from center.' }),
+    axis: PD.Select('a', PD.arrayToOptions(['a', 'b', 'c'] as const), { isEssential: true, hideIf: p => p.mode !== 'frame', description: 'Axis of the frame.' }),
+    rotation: PD.Group({
+        axis: PD.Vec3(Vec3.create(1, 0, 0), {}, { description: 'Axis of rotation' }),
+        angle: PD.Numeric(0, { min: -180, max: 180, step: 1 }, { immediateUpdate: true, description: 'Axis rotation angle in Degrees' }),
+    }, { isExpanded: true, hideIf: p => p.mode !== 'frame' }),
+};
+export type SliceParams = typeof SliceParams
+export type SliceProps = PD.Values<SliceParams>
+
+export function getSliceParams(ctx: ThemeRegistryContext, volume: Volume) {
+    const p = PD.clone(SliceParams);
+    const dim = volume.grid.cells.space.dimensions;
+    p.dimension = PD.MappedStatic('x', {
+        x: PD.Numeric(0, { min: 0, max: dim[0] - 1, step: 1 }, { immediateUpdate: true }),
+        y: PD.Numeric(0, { min: 0, max: dim[1] - 1, step: 1 }, { immediateUpdate: true }),
+        z: PD.Numeric(0, { min: 0, max: dim[2] - 1, step: 1 }, { immediateUpdate: true }),
+    }, { isEssential: true, hideIf: p => p.mode !== 'grid' });
+    p.isoValue = Volume.createIsoValueParam(Volume.IsoValue.absolute(volume.grid.stats.min), volume.grid.stats);
+    return p;
+}
+
+export async function createImage(ctx: VisualContext, volume: Volume, key: number, theme: Theme, props: SliceProps, image?: Image): Promise<Image> {
+    if (props.mode === 'frame') {
+        return createFrameImage(ctx, volume, key, theme, props, image);
+    } else {
+        return createGridImage(ctx, volume, key, theme, props, image);
+    }
+}
+
+//
+
+async function createFrameImage(ctx: VisualContext, volume: Volume, key: number, theme: Theme, props: SliceProps, image?: Image): Promise<Image> {
+    const { offset, isoValue } = props;
+
+    const { cells: { space }, stats } = volume.grid;
+    const { min, max } = stats;
+
+    const isUniform = theme.color.granularity === 'uniform';
+    const color = 'color' in theme.color && theme.color.color
+        ? theme.color.color
+        : () => Color(0xffffff);
+
+    const { size, major, minor, normal, center, trim, resolution } = getFrame(volume, props);
+    const scaleFactor = 1 / resolution;
+
+    const scale = Vec3.create(size[0], size[1], 1);
+    const offsetDir = Vec3.setMagnitude(Vec3(), normal, size[2] / 2);
+
+    const width = Math.floor(size[1] * scaleFactor);
+    const height = Math.floor(size[0] * scaleFactor);
+
+    const m = Mat4.identity();
+    const v = Vec3();
+    const anchor = Vec3();
+
+    Vec3.add(v, center, major);
+    Mat4.targetTo(m, center, v, minor);
+    Vec3.scaleAndAdd(anchor, center, offsetDir, offset);
+    Mat4.setTranslation(m, anchor);
+    Mat4.mul(m, m, Mat4.rotY90);
+    Mat4.scale(m, m, scale);
+
+    const imageArray = new Uint8Array(width * height * 4);
+    const groupArray = new Uint8Array(width * height * 4);
+    const valueArray = new Float32Array(width * height);
+
+    const [mx, my, mz] = space.dimensions;
+    const gridToCartn = Grid.getGridToCartesianTransform(volume.grid);
+    const cartnToGrid = Mat4.invert(Mat4(), gridToCartn);
+    const gridCoords = Vec3();
+
+    const pl = PositionLocation(Vec3(), Vec3());
+    const getTrilinearlyInterpolated = Grid.makeGetTrilinearlyInterpolated(volume.grid, 'none');
+
+    let i = 0;
+    for (let ih = 0; ih < height; ++ih) {
+        for (let iw = 0; iw < width; ++iw) {
+            const y = (clamp(iw + 0.5, 0, width - 1) / width) - 0.5;
+            const x = (clamp(ih + 0.5, 0, height - 1) / height) - 0.5;
+            Vec3.set(v, x, -y, 0);
+            Vec3.transformMat4(v, v, m);
+
+            Vec3.copy(gridCoords, v);
+            Vec3.transformMat4(gridCoords, gridCoords, cartnToGrid);
+
+            const ix = Math.trunc(gridCoords[0]);
+            const iy = Math.trunc(gridCoords[1]);
+            const iz = Math.trunc(gridCoords[2]);
+
+            Vec3.copy(pl.position, v);
+            const c = color(pl, false);
+            Color.toArray(c, imageArray, i);
+
+            const val = normalize(getTrilinearlyInterpolated(v), min, max);
+            if (isUniform) {
+                imageArray[i] *= val;
+                imageArray[i + 1] *= val;
+                imageArray[i + 2] *= val;
+            }
+            valueArray[i / 4] = val;
+
+            if (ix >= 0 && ix < mx && iy >= 0 && iy < my && iz >= 0 && iz < mz) {
+                packIntToRGBArray(space.dataOffset(ix, iy, iz), groupArray, i);
+            }
+
+            i += 4;
+        }
+    }
+
+    const imageTexture = { width, height, array: imageArray, flipY: true };
+    const groupTexture = { width, height, array: groupArray, flipY: true };
+    const valueTexture = { width, height, array: valueArray, flipY: true };
+
+    const corners = new Float32Array([
+        -0.5, 0.5, 0,
+        0.5, 0.5, 0,
+        -0.5, -0.5, 0,
+        0.5, -0.5, 0
+    ]);
+    transformPositionArray(m, corners, 0, 4);
+
+    const isoLevel = clamp(normalize(Volume.IsoValue.toAbsolute(isoValue, stats).absoluteValue, min, max), 0, 1);
+
+    return Image.create(imageTexture, corners, groupTexture, valueTexture, trim, isoLevel, image);
+}
+
+function getFrame(volume: Volume, props: SliceProps) {
+    const { axis, rotation, mode } = props;
+
+    const gridToCartn = Grid.getGridToCartesianTransform(volume.grid);
+    const cartnToGrid = Mat4.invert(Mat4(), gridToCartn);
+    const [nx, ny, nz] = volume.grid.cells.space.dimensions;
+
+    const a = nx - 1;
+    const b = ny - 1;
+    const c = nz - 1;
+
+    const dirA = Vec3.create(a, 0, 0);
+    const dirB = Vec3.create(0, b, 0);
+    const dirC = Vec3.create(0, 0, c);
+
+    const resolution = Math.max(a, b, c) / Math.max(nx, ny, nz);
+
+    const min = Vec3.create(0, 0, 0);
+    const max = Vec3.create(a, b, c);
+    Vec3.transformMat4(min, min, gridToCartn);
+    Vec3.transformMat4(max, max, gridToCartn);
+    const center = Vec3.center(Vec3(), max, min);
+
+    const size = Vec3();
+    const major = Vec3();
+    const minor = Vec3();
+    const normal = Vec3();
+
+    if (axis === 'c') {
+        Vec3.set(size, a, b, c);
+        Vec3.copy(major, dirA);
+        Vec3.copy(minor, dirB);
+        Vec3.copy(normal, dirC);
+    } else if (axis === 'b') {
+        Vec3.set(size, a, c, b);
+        Vec3.copy(major, dirA);
+        Vec3.copy(normal, dirB);
+        Vec3.copy(minor, dirC);
+    } else {
+        Vec3.set(size, b, c, a);
+        Vec3.copy(normal, dirA);
+        Vec3.copy(major, dirB);
+        Vec3.copy(minor, dirC);
+    }
+
+    if (rotation.angle !== 0) {
+        const ra = Vec3();
+        Vec3.scaleAndAdd(ra, ra, dirA, rotation.axis[0]);
+        Vec3.scaleAndAdd(ra, ra, dirB, rotation.axis[1]);
+        Vec3.scaleAndAdd(ra, ra, dirC, rotation.axis[2]);
+        Vec3.normalize(ra, ra);
+
+        const rm = Mat4.fromRotation(Mat4(), degToRad(rotation.angle), ra);
+        Vec3.transformDirection(major, major, rm);
+        Vec3.transformDirection(minor, minor, rm);
+        Vec3.transformDirection(normal, normal, rm);
+    }
+
+    if (mode === 'frame') {
+        const r = Vec3.distance(min, max);
+        const s = Vec3.distance(min, max) * Math.SQRT2;
+        Vec3.set(size, s, s, r);
+    }
+
+    Vec3.transformDirection(major, major, gridToCartn);
+    Vec3.transformDirection(minor, minor, gridToCartn);
+    Vec3.transformDirection(normal, normal, gridToCartn);
+
+    const trim: Image.Trim = {
+        type: 3,
+        center: Vec3.create(a / 2, b / 2, c / 2),
+        scale: Vec3.create(a, b, c),
+        rotation: Quat.identity(),
+        transform: cartnToGrid,
+    };
+
+    return { size, major, minor, normal, center, trim, resolution };
+}
+
+//
+
+async function createGridImage(ctx: VisualContext, volume: Volume, key: number, theme: Theme, props: SliceProps, image?: Image): Promise<Image> {
     const { dimension: { name: dim }, isoValue } = props;
 
     const { cells: { space, data }, stats } = volume.grid;
     const { min, max } = stats;
-    const isoVal = Volume.IsoValue.toAbsolute(isoValue, stats).absoluteValue;
 
     const isUniform = theme.color.granularity === 'uniform';
     const color = 'color' in theme.color && theme.color.color
@@ -53,36 +272,10 @@ export async function createImage(ctx: VisualContext, volume: Volume, key: numbe
 
     const imageArray = new Uint8Array(width * height * 4);
     const groupArray = getPackedGroupArray(volume.grid, props);
+    const valueArray = new Float32Array(width * height);
 
     const gridToCartn = Grid.getGridToCartesianTransform(volume.grid);
     const l = PositionLocation(Vec3(), Vec3());
-
-    let v: (ix: number, iy: number, iz: number) => number;
-    if (equalEps(isoVal, stats.min, 0.001)) {
-        v = (_ix, _iy, _iz) => 255;
-    } else if (equalEps(isoVal, stats.max, 0.001)) {
-        v = (_ix, _iy, _iz) => 0;
-    } else {
-        v = (ix, iy, iz) => {
-            return (
-                (space.get(data, ix, iy, iz) >= isoVal ? 1 : 0) * 2 +
-                (space.get(data, ix + 1, iy, iz) >= isoVal ? 1 : 0) +
-                (space.get(data, ix - 1, iy, iz) >= isoVal ? 1 : 0) +
-                (space.get(data, ix, iy + 1, iz) >= isoVal ? 1 : 0) +
-                (space.get(data, ix, iy - 1, iz) >= isoVal ? 1 : 0) +
-                (space.get(data, ix, iy, iz + 1) >= isoVal ? 1 : 0) +
-                (space.get(data, ix, iy, iz - 1) >= isoVal ? 1 : 0) +
-                (space.get(data, ix + 1, iy + 1, iz + 1) >= isoVal ? 1 : 0) +
-                (space.get(data, ix - 1, iy + 1, iz + 1) >= isoVal ? 1 : 0) +
-                (space.get(data, ix + 1, iy - 1, iz + 1) >= isoVal ? 1 : 0) +
-                (space.get(data, ix + 1, iy + 1, iz - 1) >= isoVal ? 1 : 0) +
-                (space.get(data, ix - 1, iy - 1, iz + 1) >= isoVal ? 1 : 0) +
-                (space.get(data, ix - 1, iy + 1, iz - 1) >= isoVal ? 1 : 0) +
-                (space.get(data, ix + 1, iy - 1, iz - 1) >= isoVal ? 1 : 0) +
-                (space.get(data, ix - 1, iy - 1, iz - 1) >= isoVal ? 1 : 0)
-            ) / 16 * 255;
-        };
-    }
 
     let i = 0;
     for (let iy = y0; iy < ny; ++iy) {
@@ -92,14 +285,13 @@ export async function createImage(ctx: VisualContext, volume: Volume, key: numbe
                 Vec3.transformMat4(l.position, l.position, gridToCartn);
                 Color.toArray(color(l, false), imageArray, i);
 
+                const val = normalize(space.get(data, ix, iy, iz), min, max);
                 if (isUniform) {
-                    const val = space.get(data, ix, iy, iz);
-                    const normVal = (val - min) / (max - min);
-                    imageArray[i] *= normVal * 2;
-                    imageArray[i + 1] *= normVal * 2;
-                    imageArray[i + 2] *= normVal * 2;
+                    imageArray[i] *= val;
+                    imageArray[i + 1] *= val;
+                    imageArray[i + 2] *= val;
                 }
-                imageArray[i + 3] = v(ix, iy, iz);
+                valueArray[i / 4] = val;
 
                 i += 4;
             }
@@ -108,14 +300,18 @@ export async function createImage(ctx: VisualContext, volume: Volume, key: numbe
 
     const imageTexture = { width, height, array: imageArray, flipY: true };
     const groupTexture = { width, height, array: groupArray, flipY: true };
+    const valueTexture = { width, height, array: valueArray, flipY: true };
 
     const transform = Grid.getGridToCartesianTransform(volume.grid);
     transformPositionArray(transform, corners, 0, 4);
 
-    return Image.create(imageTexture, corners, groupTexture, image);
+    const trim = Image.createEmptyTrim();
+    const isoLevel = clamp(normalize(Volume.IsoValue.toAbsolute(isoValue, stats).absoluteValue, min, max), 0, 1);
+
+    return Image.create(imageTexture, corners, groupTexture, valueTexture, trim, isoLevel, image);
 }
 
-function getSliceInfo(grid: Grid, props: PD.Values<SliceParams>) {
+function getSliceInfo(grid: Grid, props: SliceProps) {
     const { dimension: { name: dim, params: index } } = props;
     const { space } = grid.cells;
 
@@ -145,7 +341,7 @@ function getSliceInfo(grid: Grid, props: PD.Values<SliceParams>) {
     };
 }
 
-function getPackedGroupArray(grid: Grid, props: PD.Values<SliceParams>) {
+function getPackedGroupArray(grid: Grid, props: SliceProps) {
     const { space } = grid.cells;
     const { width, height, x0, y0, z0, nx, ny, nz } = getSliceInfo(grid, props);
     const groupArray = new Uint8Array(width * height * 4);
@@ -162,7 +358,7 @@ function getPackedGroupArray(grid: Grid, props: PD.Values<SliceParams>) {
     return groupArray;
 }
 
-function getGroupArray(grid: Grid, props: PD.Values<SliceParams>) {
+function getGroupArray(grid: Grid, props: SliceProps) {
     const { space } = grid.cells;
     const { width, height, x0, y0, z0, nx, ny, nz } = getSliceInfo(grid, props);
     const groupArray = new Uint32Array(width * height);
@@ -179,13 +375,18 @@ function getGroupArray(grid: Grid, props: PD.Values<SliceParams>) {
     return groupArray;
 }
 
-function getLoci(volume: Volume, props: PD.Values<SliceParams>) {
-    // TODO cache somehow?
-    const groupArray = getGroupArray(volume.grid, props);
-    return Volume.Cell.Loci(volume, SortedArray.ofUnsortedArray(groupArray));
+function getLoci(volume: Volume, props: SliceProps) {
+    // TODO: cache somehow?
+    if (props.mode === 'grid') {
+        const groupArray = getGroupArray(volume.grid, props);
+        return Volume.Cell.Loci(volume, SortedArray.ofUnsortedArray(groupArray));
+    } else {
+        // TODO: get exact groups
+        return Volume.Loci(volume);
+    }
 }
 
-function getSliceLoci(pickingId: PickingId, volume: Volume, key: number, props: PD.Values<SliceParams>, id: number) {
+function getSliceLoci(pickingId: PickingId, volume: Volume, key: number, props: SliceProps, id: number) {
     const { objectId, groupId } = pickingId;
     if (id === objectId) {
         const granularity = Volume.PickingGranularity.get(volume);
@@ -200,34 +401,11 @@ function getSliceLoci(pickingId: PickingId, volume: Volume, key: number, props: 
     return EmptyLoci;
 }
 
-function eachSlice(loci: Loci, volume: Volume, key: number, props: PD.Values<SliceParams>, apply: (interval: Interval) => boolean) {
+function eachSlice(loci: Loci, volume: Volume, key: number, props: SliceProps, apply: (interval: Interval) => boolean) {
     return eachVolumeLoci(loci, volume, undefined, apply);
 }
 
 //
-
-export const SliceParams = {
-    ...Image.Params,
-    quality: { ...Image.Params.quality, isEssential: false },
-    dimension: PD.MappedStatic('x', {
-        x: PD.Numeric(0, { min: 0, max: 0, step: 1 }),
-        y: PD.Numeric(0, { min: 0, max: 0, step: 1 }),
-        z: PD.Numeric(0, { min: 0, max: 0, step: 1 }),
-    }, { isEssential: true }),
-    isoValue: Volume.IsoValueParam,
-};
-export type SliceParams = typeof SliceParams
-export function getSliceParams(ctx: ThemeRegistryContext, volume: Volume) {
-    const p = PD.clone(SliceParams);
-    const dim = volume.grid.cells.space.dimensions;
-    p.dimension = PD.MappedStatic('x', {
-        x: PD.Numeric(0, { min: 0, max: dim[0] - 1, step: 1 }),
-        y: PD.Numeric(0, { min: 0, max: dim[1] - 1, step: 1 }),
-        z: PD.Numeric(0, { min: 0, max: dim[2] - 1, step: 1 }),
-    }, { isEssential: true });
-    p.isoValue = Volume.createIsoValueParam(Volume.IsoValue.absolute(volume.grid.stats.min), volume.grid.stats);
-    return p;
-}
 
 export function SliceVisual(materialId: number): VolumeVisual<SliceParams> {
     return VolumeVisual<Image, SliceParams>({
@@ -236,30 +414,22 @@ export function SliceVisual(materialId: number): VolumeVisual<SliceParams> {
         createLocationIterator: (volume: Volume) => LocationIterator(volume.grid.cells.data.length, 1, 1, () => NullLocation),
         getLoci: getSliceLoci,
         eachLocation: eachSlice,
-        setUpdateState: (state: VisualUpdateState, volume: Volume, newProps: PD.Values<SliceParams>, currentProps: PD.Values<SliceParams>, newTheme: Theme, currentTheme: Theme) => {
+        setUpdateState: (state: VisualUpdateState, volume: Volume, newProps: SliceProps, currentProps: SliceProps, newTheme: Theme, currentTheme: Theme) => {
             state.createGeometry = (
                 newProps.dimension.name !== currentProps.dimension.name ||
                 newProps.dimension.params !== currentProps.dimension.params ||
+                newProps.mode !== currentProps.mode ||
+                !Vec3.equals(newProps.rotation.axis, currentProps.rotation.axis) ||
+                newProps.rotation.angle !== currentProps.rotation.angle ||
+                newProps.offset !== currentProps.offset ||
+                newProps.axis !== currentProps.axis ||
                 !Volume.IsoValue.areSame(newProps.isoValue, currentProps.isoValue, volume.grid.stats) ||
                 !ColorTheme.areEqual(newTheme.color, currentTheme.color)
             );
         },
-        geometryUtils: {
-            ...Image.Utils,
-            createRenderableState: (props: PD.Values<SliceParams>) => {
-                const state = Image.Utils.createRenderableState(props);
-                updateRenderableState(state, props);
-                return state;
-            },
-            updateRenderableState
-        }
-    }, materialId);
-}
 
-function updateRenderableState(state: RenderableState, props: PD.Values<SliceParams>) {
-    Image.Utils.updateRenderableState(state, props);
-    state.opaque = false;
-    state.writeDepth = true;
+        geometryUtils: Image.Utils
+    }, materialId);
 }
 
 export function SliceRepresentation(ctx: RepresentationContext, getParams: RepresentationParamsGetter<Volume, SliceParams>): VolumeRepresentation<SliceParams> {